refactor: move test to a __tests__ folder
